requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.9.0
pydantic-settings>=2.6.0
google-cloud-aiplatform[preview]
google-cloud-pubsub>=2.26.0
google-cloud-storage>=2.14.0
python-multipart
httpx
streamlit
regex
